First commit of full working project.
Flavours were programmable in original design. Now they are simple
configurations. Configurations are good enough for now. Can add a lot
more configurations to optimize this.
Add more configurations for Windows PC cases and update to support Linux
and Android.
UI needs to be improved or replaced with a simpler one.